+++ fabric-api-dev/src/main/java/net/fabricmc/fabric/mixin/dev/BlockBoxMixin.java
@@ -17,17 +17,18 @@
 package net.fabricmc.fabric.mixin.dev;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
+
+import net.fabricmc.fabric.FabricDevProperties;
+
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 
 import net.minecraft.util.math.BlockBox;
 
-import net.fabricmc.fabric.FabricDev;
-
 @Mixin(BlockBox.class)
 public class BlockBoxMixin {
 	@ModifyExpressionValue(method = "<init>(IIIIII)V", at = @At(value = "FIELD", target = "Lnet/minecraft/SharedConstants;isDevelopment:Z"))
 	private static boolean mevIsDevelopmentForDevModule(boolean original) {
-		return original || FabricDev.THROW_ON_INVALID_BLOCK_BOXES;
+		return original || FabricDevProperties.THROW_ON_INVALID_BLOCK_BOXES;
 	}
 }
