+++ fabric-api-dev/src/main/java/net/fabricmc/fabric/mixin/dev/CommandManagerMixin.java
@@ -17,22 +17,23 @@
 package net.fabricmc.fabric.mixin.dev;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
+
+import net.fabricmc.fabric.FabricDevProperties;
+
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 
 import net.minecraft.server.command.CommandManager;
 
-import net.fabricmc.fabric.FabricDev;
-
 @Mixin(CommandManager.class)
 public class CommandManagerMixin {
 	@ModifyExpressionValue(method = "<init>", at = @At(value = "FIELD", target = "Lnet/minecraft/SharedConstants;isDevelopment:Z"))
 	private static boolean mevIsDevelopmentForDevModule(boolean original) {
-		return original || FabricDev.REGISTER_DEBUG_COMMANDS;
+		return original || FabricDevProperties.REGISTER_DEBUG_COMMANDS;
 	}
 
 	@ModifyExpressionValue(method = "execute", at = @At(value = "FIELD", target = "Lnet/minecraft/SharedConstants;isDevelopment:Z"))
 	private static boolean mevIsDevelopmentForDevModule2(boolean original) {
-		return original || FabricDev.ENABLE_COMMAND_EXCEPTION_LOGGING;
+		return original || FabricDevProperties.ENABLE_COMMAND_EXCEPTION_LOGGING;
 	}
 }
