+++ fabric-api-dev/src/main/java/net/fabricmc/fabric/mixin/dev/WeightMixin.java
@@ -22,12 +22,12 @@ import org.spongepowered.asm.mixin.injection.At;
 
 import net.minecraft.util.collection.Weight;
 
-import net.fabricmc.fabric.FabricDev;
+import net.fabricmc.fabric.FabricDevProperties;
 
 @Mixin(Weight.class)
 public class WeightMixin {
 	@ModifyExpressionValue(method = "validate", at = @At(value = "FIELD", target = "Lnet/minecraft/SharedConstants;isDevelopment:Z"))
 	private static boolean mevIsDevelopmentForDevModule(boolean original) {
-		return original || FabricDev.ZERO_WEIGHT_WARNING;
+		return original || FabricDevProperties.ZERO_WEIGHT_WARNING;
 	}
 }
