+++ gradle.properties
@@ -13,7 +13,7 @@ curseforge_minecraft_version=1.21.1
 
 # Do not manually update, use the bumpversions task:
 fabric-api-base-version=0.4.42
-fabric-api-dev-version=0.0.1
+fabric-api-dev-version=1.0.0
 fabric-api-lookup-api-v1-version=1.6.68
 fabric-biome-api-v1-version=13.0.29
 fabric-block-api-v1-version=1.0.22
